pyproject.toml
[project]
name = "llm-knowledge-base"
version = "0.1.0"
description = "LLM-compiled knowledge base wiki with Q&A and search"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.83.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "flask>=3.0",
    "rank-bm25>=0.2",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "matplotlib>=3.8",
    "pymupdf>=1.24",
    "python-frontmatter>=1.1",
]

[project.scripts]
kb = "src.cli:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
